cleanup local tests
